Error Handling

+++ common/src/main/java/ps/exalt/shopping/app/common/service/impl/MongoBaseServiceImpl.java
@@ -1,19 +1,27 @@
+//////////////////////////////////////////////////
+////          author: Nour
+////          filename: MongoBaseServiceImpl.java
+////          2023
+//////////////////////////////////////////////////
 package ps.exalt.shopping.app.common.service.impl;
 
+import lombok.SneakyThrows;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Service;
 import ps.exalt.shopping.app.common.dto.BaseRequest;
 import ps.exalt.shopping.app.common.dto.BaseResponse;
+import ps.exalt.shopping.app.common.error.exception.OperationFailedException;
 import ps.exalt.shopping.app.common.model.BaseModel;
 import ps.exalt.shopping.app.common.service.BaseService;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
+
 @Service
 public abstract class MongoBaseServiceImpl<RQ extends BaseRequest<K>,
         M extends BaseModel<K>, RS extends BaseResponse, K>
         implements BaseService<RQ, M, RS, K> {
-
 
     public abstract M requestToModel(RQ rq);
 
@@ -25,13 +33,21 @@ public abstract class MongoBaseServiceImpl<RQ extends BaseRequest<K>,
 
     public abstract MongoRepository<M, K> getMongoRepository();
 
+    @SneakyThrows
     public RS create(RQ rq) {
-        M m = requestToModel(rq);
-        return modelToResponse(getMongoRepository().save(m));
+        M model = requestToModel(rq);
+        return modelToResponse(getMongoRepository().save(model));
     }
 
+    @SneakyThrows
     public void delete(K k) {
-        getMongoRepository().deleteById(k);
+        Optional<M> m = getMongoRepository().findById(k);
+        if (m.isPresent()) {
+
+            getMongoRepository().deleteById(k);
+        } else {
+            throw OperationFailedException.createOperationFailedException(getResourceBundle(), "COMMON_00002", k);
+        }
     }
 
     @Override
@@ -41,21 +57,26 @@ public abstract class MongoBaseServiceImpl<RQ extends BaseRequest<K>,
         return modelToResponse(baseModels);
     }
 
+    @SneakyThrows
     @Override
     public RS read(K k) {
-        return modelToResponse(getMongoRepository().findById(k).orElseThrow());
+
+        Optional<M> m = getMongoRepository().findById(k);
+        if (m.isPresent()) {
+            return modelToResponse(getMongoRepository().findById(k).get());
+        } else {
+            throw OperationFailedException.createOperationFailedException(getResourceBundle(), "COMMON_00002", k);
+        }
     }
 
     @Override
     public void update(RQ rq) {
-        if (getMongoRepository().existsById(rq.getId())) {
-            M oldModel = getMongoRepository().findById(rq.getId()).orElseThrow();
-            M newModel = requestToModel(rq);
-            newModel.setCreationTime(oldModel.getCreationTime());
-            getMongoRepository().save(newModel);
-        }
-
+        M oldModel = getMongoRepository().findById(rq.getId()).orElseThrow();
+        M newModel = requestToModel(rq);
+        newModel.setCreationTime(oldModel.getCreationTime());
+        getMongoRepository().save(newModel);
     }
+
     @Override
     public boolean idExists(K k) {
         return getMongoRepository().existsById(k);
